requirements.txt
black==23.7.0
click==8.1.4
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
fonttools==4.41.0
google-pasta==0.2.0
grpcio==1.56.2
grpcio-tools==1.56.2
kiwisolver==1.4.4
matplotlib==3.7.2
mypy-extensions==1.0.0
numpy==1.25.1
packaging==23.1
pathspec==0.11.1
Pillow==10.0.0
platformdirs==3.8.1
protobuf==4.23.4
pyparsing==3.0.9
python-dateutil==2.8.2
scipy==1.11.1
six==1.16.0
tomli==2.0.1
